[Build System: CMake] Move the Compatibility50 library into the stdlib/toolchain source directory.
