+++ stdlib/public/CMakeLists.txt
@@ -61,7 +61,6 @@ if(SWIFT_BUILD_STDLIB)
   add_subdirectory(stubs)
   add_subdirectory(core)
   add_subdirectory(SwiftOnoneSupport)
-  add_subdirectory(Compatibility50)
   add_subdirectory(CompatibilityDynamicReplacements)
 endif()
